requirements.txt
flask
plotly

flask_wtf
email-validator
weasyprint

biopython
primer3-py
flask-cors

openai

yfinance

python-dotenv
flask-cors
primer3-py
viennarna
plotly
pandas
yfinance
Pillow
weasyprint
wordcloud
matplotlib

